Have every asset record include the asset ID
Rather than, for each asset, having the asset ID only saved as the index
to that asset in the Dict of all assets of that type.

This should make working with and passing around the different asset
types more straightforward, as can just use the assets themselves rather
than needing to use tuples of both ID and the corresponding asset for
the common case of needing access to both.