Get initial version of network diagram drawing
Currently consists of a main line representing the full network and
connections between this and every switch and network adapter that is
part of the network.

Will only actually work at the moment for a single network under
particular circumstances (when not zoomed in too much etc.), and with no
labelling of different parts of the network; incremental improvements to
this to follow soon.