Break out modules for most types in Main
Split out Data modules for all the types representing data decoded from
the initial passed in flag, and Geometry modules for our various types
for working with geometry (to do with drawing things in the SVG layer of
the view).

Initially at least most of these modules just consist of a type
definition and decoder (where something needs to be decoded), but this
should now give us a bit more structure to the app and make things
clearer.