Display assets in deterministic order
Display all assets rendered in view in a deterministic order, rather
than the order in each case that they happen to be given by
`Dict.values`.

Currently they are just ordered by name; later we could order them by
the slot they are in (or equivalent), once we have this information
available.