Refer to just axis rather than x-axis
If anything this x-coordinate/the corresponding line is a y-axis rather
than an x-axis, since it is a vertical line parallel with the main
y-axis, so calling it an x-axis is confusing. It's just as clear though
to just call it an axis, and this means we won't need to change this
again if how exactly we're drawing networks changes.