Space out connection lines along node LHS
Similarly to what we already do for NetworkAdapters and NetworkSwitches,
draw network connection lines for a Node spaced out along the left-hand
side of the Node's DOM element, rather than all being drawn to the
centre of this side; this makes the diagram a bit nicer and easier to
read.

Unlike NetworkAdapters and NetworkSwitches, Nodes are not currently
proportionally sized to how many Network connections they have; this
does not seem necessary in this case since additional space is not
needed to display labels alongside these lines.