Stop import skipping local node
By adding a node with hostname `local` to `/etc/hosts` on the
development controller we are using, as well as adding a Chassis asset
containing the Server for the `local` Node, importing and displaying the
`local` Node now works, and so we no longer need to skip it.

See also: https://alces.slack.com/archives/CBADGPNRK/p1531932063000137.