Prevent clearing data after layout switch
Due to the use of `union` rather than `intersect`, if we received new
asset positions data from JS which doesn't include some of the current
assets then the current assets would be cleared out of the Model rather
than just retained unchanged.

In particular this now had a visible effect after switching to the
genders layout, scrolling the page so the (lack of) new positions for
the NetworkAdapters and NetworkSwitches is sent, and then switching back
to the physical layout where these should reappear, but instead all of
these and the Network connections between them would no longer be shown.