Have switch elements expand to contain OOBs
Previously we were explicitly setting the height of every switch to the
same value, so that this was sufficient for there to be enough space to
draw all network connection lines for the switch with the most network
connections. However, now that switches also each contain an OOB
element, explicitly setting the switch height was causing the OOB
element to expand outside of the switch element, as well as potentially
leading to overlap between each switch's network connection lines and
it's OOB's network connection line.

One way to prevent this would be to account for the size of the OOB when
deciding the height to set for the switch, however this would involve
logic for determining how big the switch elements should be to contain
their children, something browsers already do well.

Instead, a simpler, slightly hacky fix which works (with the current
development data at least), is to no longer explicitly set switch
heights, and just let the browser determine what height these should be,
with just a bit of extra spacing to allow room for all current network
connections to be drawn.