Find controller password only when needed
This password is required when the `alces:import` task runs, and is read
from an environment variable, however if this is done at the top-level
it will cause an exception whenever this file is required if this
environment variable is not set. Reading it from a function within the
`import` task when needed prevents this error.