Switch import task to use new `metal eval` command
And also delete commented code related to achieving the same thing in
older Metalware; this is never likely to be needed and the `metal eval`
command being available is now just the minimum requirement to import
data from a Metalware installation with Flight Inventory.

Requires the Metalware installation to be imported to include the
changes made in https://github.com/alces-software/metalware/pull/435.