Adjust defining/displaying network connections
Following the Metalware change made in
https://github.com/alces-software/metalware/pull/436, and as recently
requested, the interface of a network connection is now a property of
the NetworkConnection rather than the NetworkAdapterPort. This is more
correct as the interface used for a network connection is specific to
the Node that is connected. Additionally a port number is now associated
with each NetworkAdapterPort, which will be used to distinguish the
ports associated with a NetworkAdapter instead.

The data importing and displaying has then been adapted following these
data model changes, while displaying the interface again, this time
alongside the Node-NetworkAdapterPort connection instead, is still to be
added.