Only debounce resending positions in development
Following meeting feedback, since performance only seems degraded to the
point of being unusable in development, for now we only want to debounce
sending the messages in development.

In production this seems not quite as bad, and so we still want things
to continually update rather than debounce - opinions may change on this
however, and/or we should investigate improving performance generally so
we don't need to do this at all.